[Refactor] #223 생년월일 입력 자동 포맷 로직 공통 유틸로 분리
