+++ DOKI/Presentation/MyPage/View/Profile/MyProfileView.swift
@@ -75,7 +75,7 @@ extension MyProfileView {
             .onChange(of: viewModel.birthDay) { old, new in
                 guard new.count >= old.count else { return }
                 
-                let formatted = viewModel.formatBirthDate(new)
+                let formatted = viewModel.autoFormatBirth(new)
                 if formatted != new {
                     viewModel.birthDay = formatted
                 }
