+++ DOKI/Presentation/MyPage/ViewModel/MyProfileViewModel.swift
@@ -28,35 +28,8 @@ class MyProfileViewModel: ObservableObject {
         self.gender = gender
     }
     
-    func formatBirthDate(_ input: String) -> String {
-        let numbersOnly = input.filter { $0.isNumber }
-        let limited = String(numbersOnly.prefix(8))
-        
-        let count = limited.count
-        
-        switch count {
-        case 0...3:
-            return limited
-            
-        case 4:
-            return limited + "/"
-            
-        case 5:
-            let year = limited.prefix(4)
-            let month = limited.suffix(1)
-            return "\(year)/\(month)"
-            
-        case 6:
-            let year = limited.prefix(4)
-            let month = limited.dropFirst(4)
-            return "\(year)/\(month)/"
-            
-        default:
-            let year = limited.prefix(4)
-            let month = limited.dropFirst(4).prefix(2)
-            let day = limited.dropFirst(6)
-            return "\(year)/\(month)/\(day)"
-        }
+    func autoFormatBirth(_ input: String) -> String {
+        BirthDateInputFormatter.autoFormat(input)
     }
     
     func saveButtonTapped() {
